% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubDoublets.R
\name{callDoublets}
\alias{callDoublets}
\title{reset the doublet score threshold}
\usage{
callDoublets(doublet_scores_obs, doublet_scores_sim, expected_doublet_rate,
  doublet_score_threshold, verbose)
}
\arguments{
\item{doublet_scores_obs}{doublet scores of obs cells}

\item{doublet_scores_sim}{doublet scores of sim cells}

\item{expected_doublet_rate}{expected doublet rate}

\item{doublet_score_threshold}{reset the threshold of doublet score}

\item{verbose}{display processing messages}
}
\value{
predicted doublet scores
}
\description{
reset the doublet score threshold
}
