% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscrublet.R
\docType{data}
\name{pbmc8k}
\alias{pbmc8k}
\title{Example pbmc8k dataset}
\format{
dgTMatrix with UMI counts
}
\source{
\url{http://cf.10xgenomics.com/samples/cell-exp/2.1.0/pbmc8k/pbmc8k_filtered_gene_bc_matrices.tar.gz}
}
\usage{
pbmc8k
}
\description{
A 8k pbmc single cell dataset
}
\keyword{datasets}
