% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscrublet.R
\name{plot_doublet_histogram}
\alias{plot_doublet_histogram}
\title{Plot histogram of doublet scores for observed transcriptomes and simulated doublets}
\usage{
plot_doublet_histogram(nnc, breaks = 30)
}
\arguments{
\item{nnc}{list, output of call_doublets function}

\item{breaks}{number of breaks to be used in hist}
}
\description{
The histogram for simulated doublets is useful for determining the correct doublet
score threshold. To set threshold to a specific value, T, run call_doublets(threshold=thr.value).
}
\examples{
#' scrr = scrub_doublets(E_obs = pbmc8k,expected_doublet_rate=0.06,min_counts=2, min_cells=3, min_gene_variability_pctl=85, n_prin_comps=30)
scrr=call_doublets(scrr)
plot_doublet_histogram(scrr)
}
