% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubDoublets.R
\name{scrubDoublets}
\alias{scrubDoublets}
\title{The main function used to identify doublets}
\usage{
scrubDoublets(exp, n_neighbors = NULL, doublet_score_threshold = NULL,
  sim_doublet_ratio = 2, expected_doublet_rate = 0.06,
  stdev_doublet_rate = 0.02, synthetic_doublet_umi_subsampling = 1,
  use_approx_neighbors = TRUE, distance_metric = "euclidean",
  min_counts = 2, min_cells = 3, min_gene_variability_pctl = 85,
  log_transform = FALSE, z_score = TRUE, n_prin_comps = 30,
  verbose = TRUE)
}
\arguments{
\item{exp}{expression matrix of a single-cell RNA-seq data set (gene, cell)}

\item{n_neighbors}{the number of neighbors used to create the KNN graph}

\item{doublet_score_threshold}{the threshold used to separate doublets}

\item{sim_doublet_ratio}{the expected doublet ratio}

\item{stdev_doublet_rate}{the standard deviation of doublet rate}

\item{synthetic_doublet_umi_subsampling}{the subsampling rate while simulation}

\item{use_approx_neighbors}{whether use approx neighbors}

\item{distance_metric}{which method use measure distance of two nodes}

\item{min_counts}{minimum read count of cell}

\item{min_cells}{minimum detected cell of gene}

\item{min_gene_variability_pctl}{minumum gene variability percentage}

\item{log_transform}{whether perform log transformation}

\item{z_score}{whether transfer to z_score}

\item{n_prin_comps}{the number of prin comps used}

\item{verbose}{whether display processing message}
}
\value{
the doublet label and score for each cell
}
\description{
The main function used to identify doublets
}
